update helpers & clean code
