+++ src/helpers/cauldron/getInterest.ts
@@ -1,0 +1,12 @@
+import { BigNumber, utils } from "ethers";
+
+const INTEREST_PRECISION = 1e2;
+
+export const getInterest = (accrueInfo: any): string => {
+  const { INTEREST_PER_SECOND }: {INTEREST_PER_SECOND: BigNumber} = accrueInfo;
+  if (!INTEREST_PER_SECOND) return "0";
+
+  const seconds = BigNumber.from("316880878")
+
+  return utils.formatUnits(INTEREST_PER_SECOND.mul(INTEREST_PRECISION).div(seconds), 2);
+};
