+++ src/mixins/borrow/cauldrons.js
@@ -16,8 +16,6 @@ import abraWsGlp from "@/utils/abi/tokensAbi/abraWsGlp";
 import { getInterest } from "@/helpers/getInterest";
 import { getTotalBorrow } from "@/helpers/getTotalBorrow";
 
-import { getUserCauldronPositions } from "@/mixins/cauldron/positions";
-
 export default {
   computed: {
     ...mapGetters({
@@ -908,11 +906,6 @@ export default {
     },
   },
   async created() {
-    //this.createPools(); // TEST
-    await getUserCauldronPositions(this.chainId, this.account, this.contractProvider)
-    await getUserCauldronPositions(this.chainId, this.account, this.contractProvider)
-    await getUserCauldronPositions(this.chainId, this.account, this.contractProvider)
-    await getUserCauldronPositions(this.chainId, this.account, this.contractProvider)
-    await getUserCauldronPositions(this.chainId, this.account, this.contractProvider)
+    await this.createPools();
   },
 };
